requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.5
pluggy==1.5.0
anyio==4.8.0
langsmith==0.3.10